Added motion and calibrate buttons to dicomViewer
Added a button to create motion vectors for the ROI, but it doesn't have
any functionality yet. Also added a pixel calibration button to convert
pixels to mm; I don't really know if it is needed, but thought it could
be useful. Right now I am working on the tracking and motion functions
to try and get them to work.