Added settings button, save as dicom stack option, made properties public
Settings button allows users to view/change:
-cm/pixel calibration factor
-percentage of pixels tracked
-frames used to calculate accumulated strain.
Some additions still need to be made to the settings function make this
work properly.

With the save button the option to save the image stack as a .dcm was
added in addition to .tif.

Since many of the functions reference the same global properties, and
there was issues with some functions being able to reference them, the
properties were made public instead of private and this seems to have
fixed the issue.